style game control buttons and move rightsidebar styles into separate file
