+++ components/RightSideBar/RightSideBar.jsx
@@ -1,8 +1,9 @@
-import react from "react";
-import { View, StyleSheet } from "react-native";
+
+import { View} from "react-native";
 import MoveLog from "./MoveLog";
 import GameControls from "./GameControls";
 import ChatContainer from "../Chat/ChatContainer";
+import styles from "./RightSideBarStyles";
 
 const RightSideBar = ({
   socket,
@@ -31,15 +32,5 @@ const RightSideBar = ({
     </View>
   );
 };
-
-const styles = StyleSheet.create({
-  rightSideBar: {
-    flexDirection: "column",
-    height: "100%",
-    width: "20%",
-    justifyContent: "space-between",
-    backgroundColor: "black"
-  }
-});
 
 export default RightSideBar;
