+++ components/RightSideBar/RightSideBarStyles.js
@@ -1,0 +1,14 @@
+
+import { StyleSheet } from "react-native";
+
+const styles = StyleSheet.create({
+    rightSideBar: {
+      flexDirection: "column",
+      height: "100%",
+      width: "20%",
+      justifyContent: "space-between",
+      backgroundColor: "black"
+    }
+  });
+
+  export default styles;
